add a single processing gpu test.
