+++ torch_xla/csrc/runtime/pjrt_registry.cc
@@ -157,6 +157,9 @@ InitializePjRt(const std::string& device_type) {
     std::shared_ptr<xla::KeyValueStoreInterface> kv_store;
     std::optional<std::set<int>> allowed_devices;
     bool spmd = sys_util::GetEnvBool("XLA_USE_SPMD", false);
+    TF_VLOG(3) << "Getting StreamExecutorGpuClient for node_id="
+               << global_process_rank << ", num_nodes=" << global_world_size
+               << ", spmd case=" << spmd << ", PJRT_LOCAL_PROCESS_RANK=" << sys_util::GetEnvInt(env::kEnvPjRtLocalRank, -1) << ", RANK=" << sys_util::GetEnvInt("RANK", -1) << ", LOCAL_WORLD_SIZE=" << sys_util::GetEnvInt("LOCAL_WORLD_SIZE", -1) << ", WORLD_SIZE=" << sys_util::GetEnvInt("WORLD_SIZE", -1);
     if (local_world_size == 1) {
       if (global_world_size > 1) {
         coordinator = SetKeyValueCallback(global_process_rank,
@@ -167,9 +170,6 @@ InitializePjRt(const std::string& device_type) {
       coordinator =
           SetKeyValueCallback(global_process_rank, global_world_size, kv_store);
     }
-    TF_VLOG(3) << "Getting StreamExecutorGpuClient for node_id="
-               << global_process_rank << ", num_nodes=" << global_world_size
-               << ", spmd case=" << spmd;
 
     xla::GpuClientOptions options;
     options.allocator_config = GetGpuAllocatorConfig();
